+++ backend/src/main/java/com/Fijalkowskim/SpaceFarmManagmentSystem/services/impl/StationServiceImpl.java
@@ -2,6 +2,7 @@ package com.Fijalkowskim.SpaceFarmManagmentSystem.services.impl;
 import com.Fijalkowskim.SpaceFarmManagmentSystem.exceptions.CustomHTTPException;
 import com.Fijalkowskim.SpaceFarmManagmentSystem.models.Cultivation;
 import com.Fijalkowskim.SpaceFarmManagmentSystem.models.Station;
+import com.Fijalkowskim.SpaceFarmManagmentSystem.repositories.CultivationDAORepository;
 import com.Fijalkowskim.SpaceFarmManagmentSystem.repositories.StationDAORepository;
 import com.Fijalkowskim.SpaceFarmManagmentSystem.services.StationService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,10 +23,12 @@ import java.util.Set;
 @Service
 public class StationServiceImpl implements StationService {
     private final StationDAORepository stationDAORepository;
+    private final CultivationDAORepository cultivationDAORepository;
 
     @Autowired
-    public StationServiceImpl(StationDAORepository stationDAORepository) {
+    public StationServiceImpl(StationDAORepository stationDAORepository, CultivationDAORepository cultivationDAORepository) {
         this.stationDAORepository = stationDAORepository;
+        this.cultivationDAORepository = cultivationDAORepository;
     }
     public Page<Station> getStations(Pageable pageable){
         return stationDAORepository.findAll(pageable);
@@ -44,6 +47,10 @@ public class StationServiceImpl implements StationService {
     }
     public void deleteStation(Long id) throws CustomHTTPException {
         Optional<Station> station = stationDAORepository.findById(id);
+        Optional<Cultivation> cultivation = cultivationDAORepository.findCultivationStationById(id);
+        if(cultivation.isPresent()){
+            throw new CustomHTTPException("Station assigned to cultivation", HttpStatus.FOUND);
+        }
         if(station.isEmpty()) throw new CustomHTTPException("Station not found", HttpStatus.NOT_FOUND);
         stationDAORepository.delete(station.get());
     }
